+++ src/io-loop.c
@@ -108,7 +108,7 @@ static struct cmds_tag {
     { "time",       cmd_time,       true,  "/time <target>" },
     { "topic",      cmd_topic,      true,  "/topic [new topic]" },
     { "version",    cmd_version,    true,  "/version <target>" },
-    { "who",        cmd_who,        true,  "/who [mask]" },
+    { "who",        cmd_who,        true,  "/who <mask>" },
     { "whois",      cmd_whois,      true,  "/whois <nick>" },
 };