+++ src/network.h
@@ -7,6 +7,8 @@
 #include "net-w32.h"
 #endif
 
+#include "x509_check_host.h"
+
 struct network_connect_context {
     char *server;
     char *port;
@@ -37,5 +39,6 @@ void	net_ssl_close(void);
 int	net_ssl_start(void);
 int	net_ssl_send(const char *fmt, ...);
 int	net_ssl_recv(struct network_recv_context *ctx, char *recvbuf, int recvbuf_size);
+int	net_ssl_check_hostname(const char *, unsigned int);
 
 #endif