+++ src/config.c
@@ -60,6 +60,7 @@ static struct tagConfDefValues {
     { "connection_timeout",        TYPE_INTEGER, "30" },
     { "disable_beeps",             TYPE_BOOLEAN, "no" },
     { "encoding",                  TYPE_STRING,  "iso-8859-1" },
+    { "hostname_checking",         TYPE_BOOLEAN, "yes" },
     { "kick_close_window",         TYPE_BOOLEAN, "yes" },
     { "max_chat_windows",          TYPE_INTEGER, "60" },
     { "nickname",                  TYPE_STRING,  "warezkid" },