+++ src/io-loop.c
@@ -383,7 +383,7 @@ add_to_history(const char *string)
 void
 enter_io_loop(void)
 {
-    new_window_title(g_status_window_label, "Swirc titlebar [tm]");
+    new_window_title(g_status_window_label, g_swircWebAddr);
 
     if (config_bool_unparse("startup_greeting", true)) {
 	swirc_greeting();