+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Added ###
 - Event 461 `ERR_NEEDMOREPARAMS`
 
+### Fixed ###
+- Usage for command /who was bogus. Parameter mask is mandatory.
+
 ## [1.4] - 2017-02-03 ##
 ### Added ###
 - WIN32 specific: added a resource script