+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 ### Changed ###
 - Connection timeout to 45 seconds
 - Set `ssl_verify_peer` to YES even on WIN32
+- ASCII logo
 
 ### Fixed ###
 - Usage for command /who was bogus. Parameter mask is mandatory.