+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - Key enter for Windows
 - Unget `MY_KEY_RESIZE` right after connection to a server. May or may
   not fix junk characters from sometimes appearing at the prompt.
+- Command history bug fix
 
 ## [1.4] - 2017-02-03 ##
 ### Added ###